Separate available trips functionality into AvailableTrips component, Moved available trips state, fetching, and pick/cancel logic from StartTrip to AvailableTrips.
